Create Auth guard and modify Auth service
The purpose of this commit is to allow for development of components
that use the Auth guard and/or service. The Auth guard was created to
restrict access to certain pages by redirecting to a login page. The
Auth service, which had a deliverable login() method already, will
temporarily implement a mockLogin() method which allows the developer to
hit Submit and be granted access instantly in order to streamline
development. This will obviously need to be changed once the guarded
features are fully implemented.