Add char. limits for creating & updating comments
Now, CommentComponent has a 500 character limit, which is enforced by
disabling the submit button on a 0 or >500 character string. A character
counter is now visible under Create/Update textarea elements, which
turns red when the input length is invalid.

The CSS for textarea and button elements has been updated. This is to
allow the newly implemented character counter to fit between the two
without looking much different.

Finally, the textarea background color rule has been removed to appear
more subtle. This involved adding a border with the old background color
(#7f7f7f).