Add character limit to create comment input
First, add comment and post character limits as environment variables so
they are consistent across components.

Add a Boolean getter to evaluate if comment length is valid, just as
needed in CommentComponent. Use this in the template when evaluating
color, fill, and disabled attributes for buttons.

Add a clear button that changes state between a disabled, empty string
state and an active, non-empty string state. Also, bind textarea value
to commentInput, so clicking the clear button will update the input
field.

Move border-radius rule to parent in CommentComponent to make it DRY.