Fix duplicate blog comments bug; add end message
Before, when navigating to a blog post from the blog page, there was a
bug that caused the first page of comments to be duplicated. This was
due to a race condition, which was solved by introducing a new variable:
commentsLoading. The old variable commentsLoaded was renamed to
commentsEnded for extra clarity.

Also, when the last page of comments has been loaded, a message is
displayed stating that the comments have all loaded.